Add: tools/moment.js
