+++ src/components/common/room/Room.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useSpring, animated } from "react-spring";
 import PropTypes from "prop-types";
-import { date2str } from "tools/trans";
+import { date2str } from "tools/moment";
 import "./RoomElement.css";
 
 import ArrowRightAltRoundedIcon from "@mui/icons-material/ArrowRightAltRounded";
