+++ src/components/common/roomOptions/DatePicker.jsx
@@ -1,6 +1,6 @@
 import React, { Component, useState } from "react";
 import { useSpring, animated } from "react-spring";
-import { getToday10 } from "tools/trans";
+import { getToday10 } from "tools/moment";
 import PropTypes from "prop-types";
 
 import CalendarTodayIcon from "@material-ui/icons/CalendarToday";
